+++ faros-airbyte-common/src/azure-devops/types.ts
@@ -2,8 +2,16 @@ import {AxiosInstance} from 'axios';
 import {IBuildApi} from 'azure-devops-node-api/BuildApi';
 import {ICoreApi} from 'azure-devops-node-api/CoreApi';
 import {IGitApi} from 'azure-devops-node-api/GitApi';
+import {
+  Build as BaseBuild,
+  BuildArtifact,
+  TimelineRecord as BaseTimelineRecord,
+} from 'azure-devops-node-api/interfaces/BuildInterfaces';
 import {IdentityRef} from 'azure-devops-node-api/interfaces/common/VSSInterfaces';
 import {GraphUser} from 'azure-devops-node-api/interfaces/GraphInterfaces';
+import {Pipeline as BasePipeline} from 'azure-devops-node-api/interfaces/PipelinesInterfaces';
+import {ProjectReference} from 'azure-devops-node-api/interfaces/ReleaseInterfaces';
+import {CodeCoverageStatistics} from 'azure-devops-node-api/interfaces/TestInterfaces';
 import {IPipelinesApi} from 'azure-devops-node-api/PipelinesApi';
 import {IReleaseApi} from 'azure-devops-node-api/ReleaseApi';
 import {ITestApi} from 'azure-devops-node-api/TestApi';
@@ -45,4 +53,23 @@ export type User = GraphUser | IdentityRef;
 export interface GraphUserResponse {
   count: number;
   value: GraphUser[];
+}
+
+export interface Pipeline extends BasePipeline {
+  project: ProjectReference;
+}
+// Ensure Build reason, status, and result enums are strings
+export interface Build extends Omit<BaseBuild, 'reason' | 'status' | 'result'> {
+  artifacts: BuildArtifact[];
+  coverageStats: CodeCoverageStatistics[];
+  jobs: TimelineRecord[];
+  reason: string;
+  status: string;
+  result: string;
+}
+
+export interface TimelineRecord
+  extends Omit<BaseTimelineRecord, 'result' | 'state'> {
+  result: string;
+  state: string;
 }
