--- faros-airbyte-common/src/azurepipeline/index.ts
@@ -1,1 +1,0 @@
-export * from './types';
