+++ faros-airbyte-common/test/azure-devops/azure-devops.test.ts
@@ -151,7 +151,9 @@ describe('client', () => {
 
   test('paginate with continuation token', async () => {
     const getPage = (pageSize: number, continuationToken?: number) => {
-      const startIndex = continuationToken ? continuationToken - 1 : 0;
+      const startIndex = continuationToken
+        ? paginationItems.findIndex((item) => item.id === continuationToken)
+        : 0;
       return paginationItems.slice(startIndex, startIndex + pageSize);
     };
